// app/admin/_lib/tareaFuncion.type.ts
export type TareaFuncionNuevaFormData = {
    nombreInterno: string;
    nombreVisible: string;
    descripcion: string; // En el schema es String?, pero en el form lo haremos string vacío por defecto
    parametros: ParametroSeleccionableEnForm[];
};
// ...
export interface ParametroSeleccionableEnForm extends ParametroRequeridoBasePrisma {
    seleccionado: boolean;
    esObligatorio: boolean;
}
// ...
export type CrearTareaFuncionInput = Pick<TareaFuncionBasePrisma, 'nombreInterno' | 'nombreVisible'> &
    Partial<Pick<TareaFuncionBasePrisma, 'descripcion'>> & {
        parametros?: Array<{
            parametroRequeridoId: string;
            esObligatorio: boolean;
        }>;
    };